Upload self-profile files to S3 after each collection
